Created an RecordSetIterator Interface to front either a XmlRecordSetIterator (renamed from the old RecordSetIterator) or a JsonRecordSetIterator.  This will break existing code that tries to instantiate a RecordSetIterator.  I don't know of too many orgs that are using it.
