+++ gedcomx-model/src/test/java/org/gedcomx/util/JsonRecordSetIteratorTest.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright Intellectual Reserve, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.gedcomx.util;
 
 import junit.framework.TestCase;
@@ -18,10 +33,9 @@ import java.util.List;
 import java.util.zip.GZIPOutputStream;
 
 /**
- * User: brenthale
+ * User: Brent Hale
  * Date: 6/4/2015
  * <p/>
- * Copyright FamilySearch.org
  */
 public class JsonRecordSetIteratorTest extends TestCase {
 
@@ -118,7 +132,7 @@ public class JsonRecordSetIteratorTest extends TestCase {
     InputStream inputStream = ClassLoader.getSystemClassLoader().getResourceAsStream("gedcomx-recordset.xml");
 
     int numRecords = 0;
-    RecordSetIterator recordSetIterator = new RecordSetIterator(inputStream, false);    // This is XML specific so far.
+    RecordSetIterator recordSetIterator = new XmlRecordSetIterator(inputStream, false);    // This is XML specific so far.
 
     Gedcomx record;
     List<Gedcomx> records = new ArrayList<Gedcomx>();
