+++ gedcomx-model/src/test/java/org/gedcomx/util/JsonRecordSetWriterTest.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright Intellectual Reserve, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.gedcomx.util;
 
 import junit.framework.TestCase;
@@ -17,7 +32,6 @@ import java.util.zip.GZIPOutputStream;
  * User: brenthale
  * Date: 6/8/2015
  * <p/>
- * Copyright FamilySearch.org
  */
 public class JsonRecordSetWriterTest extends TestCase {
