+++ gedcomx-model/src/test/java/org/gedcomx/util/TestRecordSetIterator.java
@@ -1,18 +1,31 @@
+/**
+ * Copyright Intellectual Reserve, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.gedcomx.util;
 
-import junit.framework.Test;
 import junit.framework.TestCase;
-import junit.framework.TestSuite;
 import org.gedcomx.Gedcomx;
 
 import javax.xml.stream.XMLStreamException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.util.Iterator;
 
 /**
  * Class for testing the RecordIterator
+ *
  * User: Randy Wilson
  * Date: 11/21/13
  * Time: 1:46 PM
@@ -21,7 +34,7 @@ public class TestRecordSetIterator extends TestCase {
 
   public void testParser() throws IOException, XMLStreamException {
     InputStream inputStream = getClass().getClassLoader().getResourceAsStream("gedcomx-recordset.xml");
-    Iterator<Gedcomx> recordIterator = new RecordSetIterator(inputStream);
+    RecordSetIterator recordIterator = new XmlRecordSetIterator(inputStream);
     assertTrue(recordIterator.hasNext());
     Gedcomx record1 = recordIterator.next();
     Gedcomx record2 = recordIterator.next();
@@ -41,7 +54,7 @@ public class TestRecordSetIterator extends TestCase {
 
   public void testParserWithMetadata() throws IOException, XMLStreamException {
     InputStream inputStream = getClass().getClassLoader().getResourceAsStream("gedcomx-recordset2.xml");
-    RecordSetIterator recordIterator = new RecordSetIterator(inputStream);
+    RecordSetIterator recordIterator = new XmlRecordSetIterator(inputStream);
     assertTrue(recordIterator.hasNext());
     // Haven't hit it yet.
     assertNull(recordIterator.getMetadata());
@@ -70,14 +83,11 @@ public class TestRecordSetIterator extends TestCase {
 
   public void testClose() throws IOException, XMLStreamException {
     URL url = getClass().getClassLoader().getResource("gedcomx-recordset2.xml");
-    RecordSetIterator recordIterator = new RecordSetIterator(url.getFile());
+    assertNotNull(url);
+
+    RecordSetIterator recordIterator = new XmlRecordSetIterator(url.getFile());
+
     assertTrue(recordIterator.hasNext());
-    // Haven't hit it yet.
-    assertNull(recordIterator.getMetadata());
-    Gedcomx record1 = recordIterator.next();
-    // Haven't hit it yet.
-    assertNull(recordIterator.getMetadata());
-    Gedcomx record2 = recordIterator.next();
 
     // Close it before it reaches the end of the file to test close.
     recordIterator.close();
