+++ gedcomx-model/src/test/java/org/gedcomx/util/TestRecordSetWriter.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright Intellectual Reserve, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.gedcomx.util;
 
 import junit.framework.TestCase;
@@ -11,6 +26,7 @@ import java.util.zip.GZIPOutputStream;
 
 /**
  * Class for testing the RecordSetWriter and RecordSetIterator class.
+ *
  * User: Randy Wilson
  * Date: 12/4/13
  * Time: 3:39 PM
@@ -22,7 +38,7 @@ public class TestRecordSetWriter extends TestCase {
       for (boolean isGzipped : new boolean[]{false, true}) {
         InputStream inputStream = getClass().getClassLoader().getResourceAsStream("gedcomx-recordset.xml");
         int numRecords = 0;
-        RecordSetIterator recordIterator = new RecordSetIterator(inputStream, false);
+        RecordSetIterator recordIterator = new XmlRecordSetIterator(inputStream, false);
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
         OutputStream outputStream = isGzipped ? new GZIPOutputStream(bos) : bos;
         RecordSetWriter writer = new RecordSetWriter(outputStream);
@@ -60,7 +76,7 @@ public class TestRecordSetWriter extends TestCase {
         assertEquals(3, records.size());
 
         byte[] bytes = bos.toByteArray();
-        recordIterator = new RecordSetIterator(new ByteArrayInputStream(bytes), isGzipped);
+        recordIterator = new XmlRecordSetIterator(new ByteArrayInputStream(bytes), isGzipped);
         for (int i = 0; i < numRecords; i++) {
           record = recordIterator.next();
           assertNotNull(record);
